Moving nano::system out of the node library and in to the test_core library since it is a testing class. Also renaming file from testing.c/hpp to system.c/hpp to better represent what the file contains.
