+++ nano/core_test/conflicts.cpp
@@ -1,5 +1,5 @@
 #include <nano/node/election.hpp>
-#include <nano/node/testing.hpp>
+#include <nano/test_common/system.hpp>
 #include <nano/test_common/testutil.hpp>
 
 #include <gtest/gtest.h>
