+++ nano/core_test/distributed_work.cpp
@@ -1,5 +1,5 @@
 #include <nano/core_test/fakes/work_peer.hpp>
-#include <nano/node/testing.hpp>
+#include <nano/test_common/system.hpp>
 #include <nano/test_common/testutil.hpp>
 
 #include <gtest/gtest.h>
