+++ nano/core_test/frontiers_confirmation.cpp
@@ -1,5 +1,5 @@
 #include <nano/node/active_transactions.hpp>
-#include <nano/node/testing.hpp>
+#include <nano/test_common/system.hpp>
 #include <nano/test_common/testutil.hpp>
 
 #include <gtest/gtest.h>
