+++ nano/core_test/peer_container.cpp
@@ -1,4 +1,4 @@
-#include <nano/node/testing.hpp>
+#include <nano/test_common/system.hpp>
 #include <nano/test_common/testutil.hpp>
 
 #include <gtest/gtest.h>
