+++ nano/core_test/telemetry.cpp
@@ -1,6 +1,6 @@
 #include <nano/node/telemetry.hpp>
-#include <nano/node/testing.hpp>
 #include <nano/test_common/telemetry.hpp>
+#include <nano/test_common/system.hpp>
 #include <nano/test_common/testutil.hpp>
 
 #include <gtest/gtest.h>
