+++ nano/core_test/wallets.cpp
@@ -1,5 +1,5 @@
-#include <nano/node/testing.hpp>
 #include <nano/secure/versioning.hpp>
+#include <nano/test_common/system.hpp>
 #include <nano/test_common/testutil.hpp>
 
 #include <gtest/gtest.h>
