+++ nano/node/CMakeLists.txt
@@ -128,8 +128,6 @@ add_library(
   state_block_signature_verification.cpp
   telemetry.hpp
   telemetry.cpp
-  testing.hpp
-  testing.cpp
   transport/tcp.hpp
   transport/tcp.cpp
   transport/transport.hpp
