+++ nano/test_common/network.cpp
@@ -1,6 +1,6 @@
 #include <nano/node/node.hpp>
-#include <nano/node/testing.hpp>
 #include <nano/test_common/network.hpp>
+#include <nano/test_common/system.hpp>
 #include <nano/test_common/testutil.hpp>
 
 #include <gtest/gtest.h>
