+++ nano/test_common/testutil.cpp
@@ -1,5 +1,5 @@
 #include <nano/crypto_lib/random_pool.hpp>
-#include <nano/node/testing.hpp>
+#include <nano/test_common/system.hpp>
 #include <nano/test_common/testutil.hpp>
 
 #include <gtest/gtest.h>
